requirements.txt
# Reddit data collection
praw==7.7.1
requests==2.31.0

# Market data
yfinance>=0.2.40

# NLP / Sentiment
vaderSentiment==3.3.2
transformers==4.40.0
torch==2.3.0

# Data processing
pandas==2.2.2
numpy==1.26.4
pyarrow==16.0.0
fastparquet==2024.2.0

# Machine learning
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
optuna==3.6.1

# Visualisation
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.22.0

# Notebook
jupyter==1.0.0
ipykernel==6.29.4
tqdm==4.66.4

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
joblib==1.4.2
openpyxl==3.1.2
